Reverting angle, for spitter attack

+++ PP-2-March2025/Assets/Scripts/Damage.cs
@@ -9,7 +9,7 @@ public class Damage : MonoBehaviour
     [SerializeField] Rigidbody rb;
 
     [Range(1, 10)][SerializeField] int damageAmount;
-    [Range(1, 50)][SerializeField] int speed;
+    [Range(1, 50)][SerializeField] float speed;
     [Range(1, 10)][SerializeField] int destroyTime;
     [Range(1, 4)][SerializeField] float damageTime;
